Introduced getpot for the configuration, muParser for Functor from parsed expression

+++ Test_check.sh
@@ -1,11 +1,8 @@
 #!/bin/sh
 cd ./test;
-rm test.log;
-export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:../plugin_loading/lib;
-export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:../refine_binary/lib;
-export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:../interpolating_functions/lib;
-export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:../quadrature_rules/libmesh_quadrature/lib;
-export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:../quadrature_rules/sandia_quadrature/lib;
-export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:../library_bridges/libmesh_bridge/lib;
+if [ -f "test.log" ]; then
+	echo Removing previous "test.log"
+	rm test.log;
+fi
 ./bin/test;
 cd ..;
